Move methods for metric calculations from Evaluator class (#1402)
* Moved metric calculations to own file & updated tests

* Fixed typo.

* Relative import for metrics in evaluation/_base.py.
